+++ inst/test-data/specific-redcapr/delete/single-arm-four-records.R
@@ -1,0 +1,7 @@
+structure(list(record_id = c(101, 104, 106, 107, 108, 109, 110, 
+111, 112, 113, 114, 115, 116, 117, 118, 119), birth_date = structure(c(18383, 
+18475, 18536, 18567, 18597, 18628, 18659, 18687, 18718, 18748, 
+18779, 18809, 18840, 18871, 18901, 18932), class = "Date"), position = c("pg", 
+"pf", "pg", "sg", "sf", "pf", "c", "pg", "sg", "sf", "pf", "c", 
+"pg", "sg", "sf", "pf"), demographics_complete = c(2, 2, 2, 2, 
+2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2)), row.names = c(NA, -16L), class = "data.frame")