+++ NEWS.md
@@ -1,6 +1,11 @@
 Upcoming Versions
 ==========================================================
 
+### New Features
+
+* `redcap_delete()` deletes a vector of records.  (Thanks @joundso, #372, #373)
+* `redcap_arm_export()` retrieves a list of REDCap project arms.  (#375)
+
 ### Minor Enhancements
 
 * `sanitize_token()` now allows lowercase characters --in addition to uppercase characters & digits. (#347, @jmbarbone)